add start argument to make_aweek and as.aweek
This mechanism is hopefully to avoid the same ambiguity problem
that I was trying to solve with c.aweek(). You can create aweek
objects from week numbers or characters, but the resulting aweek
object should not have a different week_start attribute based on
the first element in the vector.